test-web: Move results generation into Results.cpp
Declutter main in anticipation of a new results-by-pid view.
